backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
psycopg2-binary
geoalchemy2
redis
python-jose[cryptography]
passlib[bcrypt]
python-multipart
rasterio
numpy
shapely
structlog
pydantic-settings
httpx
python-dotenv
